Updated GitHubService to solve debugging issues
Beforehand, there was no iteration of the item in "jsonItemList" leaving a null reference, therefore causing a compile error when the url was trying to be retrieved. Now with this addition of another iteration through the list, the url can be retrieved from json object and added to repoNames list. 

I also changed .ToUpper() to .ToLower() when adding in url to list. This is to avoid writing any duplicate urls in file that will occur, because the urls being read from the file were lowered when adding them to the repoNamesInFile list. When checking for duplicates the strings need to be the same upper or lowercase. I could've also just changed the lines being read from file to all uppercase and kept the .ToUpper() when adding the urls in the repoNames lilst, but I decided to go lowercase for simplicity and stylistic choices.

The last change on line 84 was to add the DecryptString function, in order to match the urls being checked for from repoNames. Without this, every string being read from file would still be encrypted, and would cause the file to write many duplicates.