Updated _ViewStart.cshtml
This will cause an error trying to load the page. It doesn't point to a file in project though it tries to. The intended file would be "~/Views/Shared/_Layout.cshtml" but instead it would try to load ~/Views/_Layout.". 

The file path has to be specified in order for this to work.